requirements.txt
pandas==1.5.2
Pillow==9.3.0
Pillow==9.5.0
plotly==5.14.1
Requests==2.30.0
streamlit==1.22.0
streamlit_lottie==0.0.3
